Creates basic server interact class with dummy data
